Drop redundant uint64 conversion in floor-gas comparison
intrinsic is already a uint64 — TxGas and TxDataZeroGas are both uint64
constants from params/protocol_params.go, so the addition stays in
uint64 without a cast.

+++ x/evm/keeper/simulate_v1_test.go
@@ -679,7 +679,7 @@ func TestValidateSimCall_FloorDataGas_BelowFloorFails_38013(t *testing.T) {
 	msg := f.baselineMsg()
 	msg.Data = data
 	msg.GasLimit = floor - 1 // > intrinsic, < floor
-	require.Greater(t, msg.GasLimit, uint64(intrinsic))
+	require.Greater(t, msg.GasLimit, intrinsic)
 
 	simErr := f.k.validateSimCall(f.ctx, f.sdb, msg, f.header, rules, f.cfg)
 	require.NotNil(t, simErr)
